Move selection list header to WorkspaceSettlementAccountPage
